fixed all testcases, now starting to write lots of printfs

+++ TestCases/ABCTests/abcTest.py
@@ -131,7 +131,7 @@ class LLVMTest(unittest.TestCase, ABC):
 
             index = file[4:-2]  # The index is used to refer to the files & other data belonging to this testfile
             file_name = f"tests/test{index}.c"
-            #print(index, file_name)  # Toggle for debug
+            # print(index, file_name)  # Toggle for debug
 
             """
             If input will be read, it needs to be retrieved
@@ -186,11 +186,15 @@ class LLVMTest(unittest.TestCase, ABC):
                 """
                 errors = str(error_buff.getvalue().splitlines())
                 expected_errors = str(error_dict.get(str(index), []))
-                # print("error", error_buff.getvalue().splitlines(), index)  # Print any errors we didn't expect
+                print("error", error_buff.getvalue().splitlines(), index)  # Print any errors we didn't expect
                 assert errors == expected_errors
 
             sys.stdout = original
             sys.stderr = original_error
+
+    def compileGCC(self, file_name):
+        return subprocess.run(f"""gcc -ansi -pedantic {file_name} -o temp/temp""",
+                                   shell=True, capture_output=True)
 
     def runC(self, file_name, inp):
         """
@@ -199,13 +203,23 @@ class LLVMTest(unittest.TestCase, ABC):
         :param inp: Any optional input
         :return:
         """
-        # Run c file using gcc
-        c_out = subprocess.run(f"""gcc -ansi -pedantic {file_name} -o temp/temp""",
-                               shell=True, capture_output=True)
-        if c_out.returncode != 0:  # Compilation failed, warn client!
-            raise Exception("Compilation Failed")
 
-        # Run the compiled code
-        c_out = subprocess.run(f" temp/./temp ; rm temp/temp", shell=True, capture_output=True, text=True, input=inp)
+        json_file = file_name[:-2] + ".json"
+        index = file_name[10:-2]
+
+        if os.path.isfile(json_file):  # If a .json file exists, we need to load the value from this file as c_out instead of the gcc compiler
+            with open(json_file, "rt") as f:
+                output_dict = json.loads(f.read())
+                c_out = subprocess.run(f"""ls""", shell=True, capture_output=True) # Create a useless c_out
+                c_out.stdout = output_dict[index]  # Modify output
+                c_out.stderr = ""  # No errors expected
+        else:
+            # Run c file using gcc
+            c_out = self.compileGCC(file_name)
+            if c_out.returncode != 0:  # Compilation failed, warn client!
+                raise Exception("Compilation Failed")
+
+            # Run the compiled code
+            c_out = subprocess.run(f" temp/./temp ; rm temp/temp", shell=True, capture_output=True, text=True, input=inp)
 
         return c_out
